feat: AbstractSecurityConfig
